// HelpDeskUI/api-client/models/Column.ts
/* tslint:disable */
/* eslint-disable */
/**
 * HelpDeskAPI
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: v1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


/**
 * 
 * @export
 */
export const Column = {
    CreatedAt: 'CreatedAt',
    FirstName: 'FirstName',
    LastName: 'LastName',
    Status: 'Status',
    UpdatedAt: 'UpdatedAt'
} as const;
export type Column = typeof Column[keyof typeof Column];


export function ColumnFromJSON(json: any): Column {
    return ColumnFromJSONTyped(json, false);
}

export function ColumnFromJSONTyped(json: any, ignoreDiscriminator: boolean): Column {
    return json as Column;
}

export function ColumnToJSON(value?: Column | null): any {
    return value as any;
}