// HelpDeskUI/api-client/models/ModelSortColumn.ts
/* tslint:disable */
/* eslint-disable */
/**
 * HelpDeskAPI
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: v1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { exists, mapValues } from '../runtime';
import type { Column } from './Column';
import {
    ColumnFromJSON,
    ColumnFromJSONTyped,
    ColumnToJSON,
} from './Column';

/**
 * 
 * @export
 * @interface ModelSortColumn
 */
export interface ModelSortColumn {
    /**
     * 
     * @type {boolean}
     * @memberof ModelSortColumn
     */
    ascending: boolean;
    /**
     * 
     * @type {Column}
     * @memberof ModelSortColumn
     */
    column: Column;
}

/**
 * Check if a given object implements the ModelSortColumn interface.
 */
export function instanceOfModelSortColumn(value: object): boolean {
    let isInstance = true;
    isInstance = isInstance && "ascending" in value;
    isInstance = isInstance && "column" in value;

    return isInstance;
}

export function ModelSortColumnFromJSON(json: any): ModelSortColumn {
    return ModelSortColumnFromJSONTyped(json, false);
}

export function ModelSortColumnFromJSONTyped(json: any, ignoreDiscriminator: boolean): ModelSortColumn {
    if ((json === undefined) || (json === null)) {
        return json;
    }
    return {
        
        'ascending': json['ascending'],
        'column': ColumnFromJSON(json['column']),
    };
}

export function ModelSortColumnToJSON(value?: ModelSortColumn | null): any {
    if (value === undefined) {
        return undefined;
    }
    if (value === null) {
        return null;
    }
    return {
        
        'ascending': value.ascending,
        'column': ColumnToJSON(value.column),
    };
}