// HelpDeskUI/api-client/models/ModelUser.ts
/* tslint:disable */
/* eslint-disable */
/**
 * HelpDeskAPI
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: v1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { exists, mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface ModelUser
 */
export interface ModelUser {
    /**
     * 
     * @type {string}
     * @memberof ModelUser
     */
    firstName: string;
    /**
     * 
     * @type {string}
     * @memberof ModelUser
     */
    id: string;
    /**
     * 
     * @type {string}
     * @memberof ModelUser
     */
    lastName: string;
}

/**
 * Check if a given object implements the ModelUser interface.
 */
export function instanceOfModelUser(value: object): boolean {
    let isInstance = true;
    isInstance = isInstance && "firstName" in value;
    isInstance = isInstance && "id" in value;
    isInstance = isInstance && "lastName" in value;

    return isInstance;
}

export function ModelUserFromJSON(json: any): ModelUser {
    return ModelUserFromJSONTyped(json, false);
}

export function ModelUserFromJSONTyped(json: any, ignoreDiscriminator: boolean): ModelUser {
    if ((json === undefined) || (json === null)) {
        return json;
    }
    return {
        
        'firstName': json['firstName'],
        'id': json['id'],
        'lastName': json['lastName'],
    };
}

export function ModelUserToJSON(value?: ModelUser | null): any {
    if (value === undefined) {
        return undefined;
    }
    if (value === null) {
        return null;
    }
    return {
        
        'firstName': value.firstName,
        'id': value.id,
        'lastName': value.lastName,
    };
}