// HelpDeskUI/api-client/models/TicketStatus.ts
/* tslint:disable */
/* eslint-disable */
/**
 * HelpDeskAPI
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: v1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


/**
 * 
 * @export
 */
export const TicketStatus = {
    Closed: 'Closed',
    InProgress: 'InProgress',
    Open: 'Open'
} as const;
export type TicketStatus = typeof TicketStatus[keyof typeof TicketStatus];


export function TicketStatusFromJSON(json: any): TicketStatus {
    return TicketStatusFromJSONTyped(json, false);
}

export function TicketStatusFromJSONTyped(json: any, ignoreDiscriminator: boolean): TicketStatus {
    return json as TicketStatus;
}

export function TicketStatusToJSON(value?: TicketStatus | null): any {
    return value as any;
}